refactor(ui): move loading and empty state methods to `HSDataLoadingVC` and update relevant view controllers
