+++ HubScout/Custom Views/ViewControllers/HSDataLoadingVC.swift
@@ -1,0 +1,47 @@
+//
+//  HSDataLoadingVC.swift
+//  HubScout
+//
+//  Created by o9tech on 13/08/2024.
+//
+
+import UIKit
+
+class HSDataLoadingVC: UIViewController {
+
+    var containerView: UIView!
+
+    func showLoadingView() {
+        containerView = UIView(frame: view.bounds)
+        view.addSubview(containerView)
+
+        containerView.backgroundColor   = .systemBackground
+        containerView.alpha             = 0
+        UIView.animate(withDuration: 0.25) { self.containerView.alpha = 0.8 }
+
+        let activityIndicator = UIActivityIndicatorView(style: .large)
+        containerView.addSubview(activityIndicator)
+        activityIndicator.translatesAutoresizingMaskIntoConstraints = false
+
+        NSLayoutConstraint.activate([
+            activityIndicator.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            activityIndicator.centerYAnchor.constraint(equalTo: view.centerYAnchor)
+        ])
+        activityIndicator.startAnimating()
+    }
+
+
+    func dismissLoadingView() {
+        DispatchQueue.main.async {
+            self.containerView.removeFromSuperview()
+            self.containerView = nil
+        }
+    }
+
+
+    func showEmptyStateView(with message: String, in view: UIView) {
+        let emptyStateView = HSEmptyStateView(message: message)
+        emptyStateView.frame = view.bounds
+        view.addSubview(emptyStateView)
+    }
+}
